+++ src/index.ts
@@ -4,5 +4,3 @@ export * as aes from './aes.js'
 export * as rsa from './rsa.js'
 export * as base64 from './base64.js'
 export { binaryConcat } from './binary.js'
-/** @deprecated - Top-level exports have been moved to the base64 module and will soon be removed */
-export { encode, decode } from './base64.js'