+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.4] 2023-02-10
+## [0.8.5] 2024-02-10
+### Removed
+- `rsa.unwrapPrivateKey`, `rsa.wrapPrivateKey`
+- Top-level base64 `encode` and `decode` exports
+
+## [0.8.4] 2024-02-10
 ### Deprecated
-- `rsa.unwrapPrivateKey` (unwrapPrivateKey missed getting flagged as deprecated in 0.8.2 and will be removed in 0.8.4)
+- `rsa.unwrapPrivateKey` (unwrapPrivateKey missed getting flagged as deprecated in 0.8.2 and will be removed in 0.8.5)
 
 ## [0.8.3] 2024-02-10
 ### Fixed